Add gerrit reporting to the irc bot.
This adds support to the irc bot for leaving a review on gerrit based
on the ES query. If it finds a match it will leave a link to the bug.
If it fails to find a match it will leave a link to the wiki page for
debugging test failures.

The unit tests depend on gerritlib change: https://review.openstack.org/#/c/47396/1