Switch queries file to YAML
It is easier for a human to read, and by virtue of not requiring
escaped quotes, easier to copy/paste into a logstash field.

When copy/pasting, the newlines won't show up in the input field.
The '>' syntax in YAML indicates folding, which causes the newline
and indentation to be turned into a single space.

Change-Id: Ibd172fd4859c055096609f31ef09222147c34cf3