don't look for things only in FAILURE logs
the queries shouldn't show up in SUCCESS runs anyway, so remove
this from the query list. This will not impact the bot finding
things, as we only search FAILURE logs when datamining the gate,
however it will let us do annalysis on the queries to ensure they
don't show up in SUCCESS runs.

Change-Id: Ib092bf9730b80545fabbe9ee5c70eccc339d2e4b