Change test_queries from logical AND to OR
Instead of passing if valid query and valid launchpad, change to pass of
valid query or valid launchpad.  We want to use OR to be extra safe
before removing queries.  With this change we will keep queries around
until there are no valid hits on logstash.openstack.org and the bug is
marked as closed in launchpad.

Change-Id: I991acb3f0f6a34366a223710c7b25d98919d3751