Better end of console log indexing detection.
* elastic_recheck/query_builder.py: Due to the way multiline events
work in logstash, the "Finished: FAILURE" line is not pushed into
elasticsearch until the next console.html file is processed or some
timeout is reached (I haven't actually confirmed that the timeout works
though). This happens because this is the last line in the file that we
do not drop. To better detect the end of indexing look for the "[SCP]
Copying console log" message which should also happen at the end of each
console log file.

A better long term solution to this problem may be to have the logstash
gearman workers insert and EOF event that can ensure each log file is
completely flushed without waiting for more data.

Change-Id: I7c51e019c774f9ba15e731aca01aaffc27de1e60