Fix E122,E126,E128 items in codebase
In the land of random cleanups, let more of the whitespace rules
back in. Also explicitly exclude E125 because of the overreach,
and leave E123 excluded because it creates some kind of odd
artifacts in the current code (possibly clean it up later).

Tox.ini adjusted with comments about the fact that what we are
ignoring is there for a reason.

Change-Id: I5636cb646d7898df71b715aa0e32a68ce279ee80