make ResultSet actually inherit from list
it turns out, I was spending way to much time to make ResultSet act
like a list, when I could have just made it inherit from list and
be done with it. It manages to remove code and work just the same.

In addition, make the __repr__ for Hit be more meaningful by using
pprint. Makes print debugging of all the datastructures actually
work like you expect.

Change-Id: Ie104d4bfc06a0875f8da85121742c053b642f8f9