implementation of FacetSet for client side nested facets
this is an implementation of facets, client side, with elastic
search results. This will let us get rid of a bunch of the
uniquify code in the graph and check_success scripts, and make
it simpler to analyze by other dimensions in web console additions.

Also make Hit implement __getitem__ for easier dynamic access of
contents. Useful for programatically accessing tags.

Change-Id: Ib63ff887eb82cff0ba00109471ee48d210fda571