Add e-r query for bug 1258682
When the build times out and this fails, there are no logs really
so we have to base this on the build timeout message in the
console log.

Note that we are essentially doing a wildcard for the timeout
value but we restrict the query based on build_name to avoid
hits on some swift jobs.

Related-Bug: #1258682

Change-Id: I0db0e08627609b44ec8ea132b980021f8d7b7b9d