Fix the e-r query for bug 1258682
The query for bug 1258682 uses wildcards and that doesn't work since
wildcard analysis is disabled by default in elasticsearch for
performance reasons.

Looking at the query again, filtering on build_name is probably not
necessary and actually limits the number of hits where we see the bug.
When removing the build_name filters, there are more hits and on
different jobs, i.e. gate-rally-py26 and gate-heat-py27 start showing
hits.

Closes-Bug: #1266833
Related-Bug: #1258682

Change-Id: Ib9c3bd05592f40d1bea8f4428e8e8fb0776cdcce