refactor bot to be based on argparse
this makes the bot based on argparse, and provides a tox job that
makes running the bot in test mode a little more sane. It also
provides for a new '-n' nocomment item so you can run the bot
and not be worried that it will report to gerrit with findings.

Change-Id: If9d6a7e72dd8d9338f2dd3283cf9a761488122de