add uncategorized failure generation code
the following adds a new tool to build a web page that will give
us a full list of all the uncategorized failures and their logs.
This gives us a todo list to start building ER queries from.

this will require an infrastructure change to run this on cron to
generate a public page.

it also brings in jinja2 for html templating, which will be useful
for future html generation jobs.

Change-Id: I5114296fbac2cde9c6b0133e2717e1cc28fb631d