Add multi-project irc support to the bot
This commit adds support for multi-project/channel irc support to the bot.
To do this you list multiple channels in the bot yaml and specify which
projects get reported to that file. When a bug is encountered that targets
the listed project the message will be reported. If 'all' is specified as
a field under projects for a channel then all projects will be reported on
that channel.

Change-Id: Ic3dd76bad94213c7152c29a99c00ed23a2c01a31