Use short build_uuids in elasticSearch queries
In addition to searching by change and patch search by the short build_uuid.
This prevents us accidentally classifying multiple builds when we classify
a failure on gerrit. This can happen in the gate queue if there is a
gate reset, or if there are multiple 'recheck bug x' on a single patch
revision in the check queue.

Change-Id: I6356a971ca250ddf5f01a9734f13d0b080a62c89