Make IRC messages provide more context
Now that we are running this on all jobs (not just tempest) we are
getting significantly more IRC messages. Add failed job name to logs to
provide more context of what job is failing. For unclassified failures
also include the queue (as a unclassified unit test failure in check queue is
much less important then one in the gate).

Change-Id: I485bf06721fa5afd102b99b26e38f12449deec7b