Add query for bug 1265495
Started showing up around 1/28, spiked on 1/30 and has been around ever
since.  You'll note that this is already a 100% fail rate but it's
restricted to check-tempest-dvsm-neutron* jobs, so we could add that to
the query to filter out other jobs if we're paranoid of this conflicting
with bug 1253896 somehow.

Related-Bug: #1265495

Change-Id: I234349076327be938acfcd740078aa52102a6104