Add query for bug 1273908
Tempest is failing on a race waiting for a server to reach 'rescue'
state when a volume is attached to it. From the n-cpu logs it looks like
the libvirt driver is doing image cache management at the time of the
timeout, but since there are no failures in the logs we have to
fingerprint the failure on the Tempest timeout error in the console log.

Related-Bug: #1273908

Change-Id: I7540d8b43e259af35723199e401f2f9d16e70c26