Delete query 1258682
Query 1258682 was a catch all for jobs that timed out. We originally
needed this because when a job timed out we wouldn't always get the log
files to analyze. This has since been fixed and now this fingerprint is
covering up other bugs. This fingerprint is getting hits from:

* gate-neutron-python-26
* gate-ceilometer-python27
* gate-ceilometer-python26
* *tripleo*

Each of those failures should be dealt accordingly instead of being
squashed into a big catch all bug like this.

Change-Id: Icde61387ff3e6d00a18d969e502b3d23cff666b3