Don't include recheck instructions when unclassified failures
If there is an unclassified failure in the check queue, we want to make
it clear to the user so they will investigate the error as its most
likely a valid failure. Also don't include recheck instructions when
unclassified failure as they shouldn't be running a recheck if there is
an unclassified failure.

With us now classifying many failures from non-voting jobs, it is common
to see classified failures and no mention of the job that legitimately
failed.

Partial revert of I52044afb4f3a1bf3f22ba4c0e8d38d76271ffc00

Change-Id: I6b471b9ab9c7f36eeed93993ea086bbc9daa56b0