Add query for nova/neutron timeout bug 1307029
There are 54 hits in the last 7 days on this across the check and
periodic-qa queues, only happens with neutron.  I was unable to find any
specific errors in the nova or neutron logs for the instance or IP in
question when this fails. Given that and it's a timeout failure, I have
to query on the failing test in the console log.

Related-Bug: #1307029

Change-Id: I8af8858b835436bb016ea810ca7f4163a5351073