add a new signature for get-pip.py fail
It turns out the get-pip.py download doesn't always fail with a
straight "can't connect to host". It often will fail just by
stalling out in the download. This is kind of hard to detect, however
a reasonable heuristic is no data found after 10s and eliminate
all the 100% complete downloads with 434k content (which is the
size of get-pip.py).

tripleo needs to be filtered out because it downloads way too much
stuff which fails, but the scripts work around it (or possibly just
fail late).

Change-Id: I09dbbe7217be999fee525b8d1d9aad76456e414c