fingerprint for race bug in swift account quotas
Because this is a negative test failure, there really isn't anything
useful outside of the test failure in the logs. This is basically
a quota not getting applied fast enough, and the actual request
passing.

Move the old bug which was just a stack trace bug, to the new bug
which includes analysis.

Change-Id: I307a95a7e67ac1b4e7f944605342e37e7903ac6d
Related-Bug: #1310597