Re-add query for bug 1251784
We used to have a change I10ca08be for this but then it was gone around
December of 2013 and removed with change I9c5d52aa.

Looks like it's back though:

http://goo.gl/icbDLV

This isn't 100% failure, but it's in the high 90s in the last 2 weeks
and showing up in the stable branches, but mostly master.

The main oddity is it's only showing up in the check queue, but it is on
several different changes.  Really spiked around 4/17 but then got
relatively quiet again around 4/19.

Related-Bug: #1251784

Change-Id: I2b85d891b17c1f14db894902c8a29f556f86ec3e