Add query for bug 1315095
This commit adds a query for bug 1315095. The real error is will be
in the failures file, because nova-network isn't running. But since
we don't archive or track that directory this query should capture
the failure because it looks for network api timeouts in the cpu log.
When we start capturing the failure file we should move to use that
for the query instead.

Related-Bug: #1315095

Change-Id: Ia67096546f1b21aedf847b44a1d438f6d4f713c2