add a tool for determining unaccounted for rechecks
this is specifically not in cmd because this is designed as a
local spot checking tool only. However it's more useful in the er
repository than on my local disk.

builds a summary report of rechecks found in gerrit in the last 2
weeks, including how many had a bug, and how many had a bug we
don't know about.

Change-Id: I6a99a3bcec527652901c8aa3639926a8e4a518d5