Add query for Neutron SSH EOFError bug 1323658
This is showing up from a common check in both test_network_basic_ops
and test_network_advanced_server_ops for public ssh connectivity that
fails.  This might be related to bug 1253896 which was the old SSH bug
in Neutron but we closed that, deleted the e-r query for it and said
we'd open new bugs and write new queries for new SSH failures, so that's
what this is.

There might be a better line to filter on in the Neutron logs, but just
based on the private IP that shows up in the console log when this
fails, I couldn't find anything.  We can tighten up the query later if
the Neutron team finds something better to fingerprint.

Related-Bug: #1323658

Change-Id: I79d4a7a00dbcbdb57ae8704f00c5871db1952623