Add query for ceilometer test_ordination_assign timeout bug 1326855
This fails in the check queue also but I didn't want to allow e-r
hitting this on the check queue in case there are changes which actually
break the test for other reasons.  I wanted to fingerprint on the test
case name and the TimeoutException from the stack trace but that wasn't
working in logstash, so another reason why I'm filtering on the gate
queue only.

Related-Bug: #1326855

Change-Id: I1d1c0a6682cdb596b4a5ce88c4fcf2f29610577b