Add query for test_volume_boot_pattern ssh timeout bug 1298472
79 hits in 7 days, check and gate, all failures, started spiking on 6/5.

This is an ssh timeout failure in the tests themselves so the
fingerprint is on the console logs.  Also, to avoid collisions with
other ssh bugs affecting the gate right now I'm fingerprinting on a
unique part of the stacktrace that would only show up for
test_volume_boot_pattern failures.

Related-Bug: #1298472

Change-Id: I2db30ad92127c996c97361941afcfe142e8374b7