Add query for bug 1332512
This commit adds a query for the race condition in the ssh timeout
tempest unit tests. The actual failure error is an assertion error
with False != True. However, since this is the general failure message
if assertTrue or assertFalse fails it would be to general to
fingerprint. So instead the test name was used instead.

Change-Id: Ic83516c39f1a0e29e961eb3c1b81257c46bb257d
Related-Bug: #1332512