Expand query for nova floating IP / sqla 0.9.5 bug 1333410
The original query was limited to nova unit tests but this also hits
Tempest so we have to expand to what shows up in the console log.

There are actually no errors in the nova service logs because nova is
simply not returning results from DB queries, so it's the client that
fails when it's doing it's assertions.

Related-Bug: #1333410

Change-Id: I4182bbc2663e3fd7f4d7a3b956de5e94a9410e27