Add query for nova libvirt connection reset snapshot bug 1334398
There are 162 hits on this in the last 2 days, check and gate.

This is different from bug 1255624 in the domain operation in progress
when libvirt is reset so this is more specific to instance snapshot.

There is a small percentage (< 1%) of successful jobs that are hit but
this is as close to 100% failure as I can find.

Related-Bug: #1334398

Change-Id: I60204af0929e833be6534bae8aedb369fbd51c8b