Add query for nova create_image race bug 1334345
The error is also showing up in Tempest runs but the hit rate on
successful jobs is too high (in the 80% range), but it's a 100% failure
in the unit test jobs.  So far I'm only seeing this in the ec2 unit test
that fails in the bug report (test_create_image_with_reboot), but could
be happening elsewhere.

Rather than filter on the test case I'm filtering on the actual
exception message which could cover other tests that go down the same
path in the compute api/manager for taking a snapshot and then starting
an instance.

Restricting this to the gate queue so we don't mistakenly report on
check queue failures for newly introduced issues still under review.

Related-Bug: #1334345

Change-Id: If8c6f4e3651ba7f081e9c12193dce1e7db2fbfef