Add query for pypi mirror infra bug 1334550
This has 327 hits in 7 days, multiple queues and jobs,
it's mostly failures but has a few hits on successful
jobs. I tried filtering success vs failure to see if there
was some obvious difference for the success hits but
couldn't find one, so risk-assuming that the low
success rate will stay low to avoid false-hits.

Change-Id: I2ef8fedec07d3a62020d272ab4b0347257a2e5ef
Related-Bug: #1334550