Add query for neutron paramiko EOFError bug 1295808
Since this is an ssh validation bug in a Tempest scenario
test we're querying on a stack trace in Tempest, which
is not great but the EOFError in tags:console is too open
and there aren't any specific errors in q-svc logs (there
are some errors but they look unrelated, i.e. results of
negative tests).

32 hits in 7 days, all failures, all in check queue but
on multiple changes.

Change-Id: Ia180ec2d93ff0a7fed43c4b96ad6dc80c54985fe
Related-Bug: #1295808