Add fingerprint for bug 1341507
Bug 1341507 is persistent failure in swift-dsvm-functional, so running a
recheck on this failure won't help until the bug is fixed. Because of
that this patch shouldn't be merged until the bug is fixed, so we don't
tell people to run pointless rechecks.

Add this fingerprint primarily for bug classification purposes.

Change-Id: I41212839cb6a48995015798b7b3f13b7f8544950
Related-Bug: #1341507