e-r query for intermittent failure in glance tests
10/10 failures in last 7 days

In the bug report there are links to the relevant time periods in
the g-api and g-reg logs but the associated strings are not as
useful as that found in the console.log (as they involve a lot
of uuids that are specific only to that test run).

    https://bugs.launchpad.net/glance/+bug/1347692

As the rules state, this is not ideal, but it resulted in a
clean and accurate match for failtures in the last 7 days.

Change-Id: I34d79a0b0a3a70bb48bbeb61d32683b4585d67d9
Related-Bug: #1347692