Fix the coverage job in tox
This commit fixes the coverage job in tox. It relies on pbr to
actually make the coverage calls before and after running the tests.
However, by default pbr assumes the project name in setup.cfg is the
same as the directory containing the source. However that isn't the
case here. This commits corrects this by using the flag to pretend
that package name is the same as the source dir. This change also
sets the correct regex for testr to run the unit tests only, which is
what is needed for the coverage job. The pbr requirement is also
updated to be >=0.8.1 because this is the min version required to
use the option exploited by this patch.

Change-Id: Icea30575538571fe842632901e994be3ac0bbc80