Add query for nova shelve unit tests bug 1353654
This could be related to a few other race bugs in
how the compute service is used in these integration
type tests.

The fingerprint here is on the one line in the console
log for the mox UnexpectedMethodCallError since the
test expects the instance to be shelved and instead
it's in error state due to an unexpected
InstanceNotFound error.

If you go back 10 days on this, there are hits in the
gate, but only 7 days shows only check queue failures,
but there are a couple and at least one is totally
unrelated to where this code would be hit.

Change-Id: I4ddb84cc40ac93b2861edd02ce493e507160788e
Related-Bug: #1353654