Add query for py26/centos6 node setup bug 1355907
13 hits in 7 days, master and stable/havana, multiple
changes on the check queue, all failures.

This only shows up in gate-*-python26 jobs on the
bare-centos6-rax* nodes.  I could have filtered
on build_name or build_node but would have to use
wildcards to do so, which we'd like to avoid in
queries due to performance issues with wildcards.

Change-Id: I997f865bd7f277dc949915cf3ebe287e5dcf4c85
Related-Bug: #1355907