Add query for test_postgresql_opportunistically bug 1355929
3 hits in 7 days, all failures, multiple changes in the
check queue.

This is limited to stable/havana and the py26 job since
that's only what it hit on.  This might have been a fluke
in infra since it's only hitting on 2 changes in the check
queue and only in the py26 unit test job on stable/havana,
but could also be a latent race bug in the pg tests in
stable/havana.

Change-Id: I595369a0025d27a2674c8b2b41467b302e3688f3
Related-Bug: #1355929