Add query for neutron subnet delete race bug 1357055
I found this from the uncategorized bugs page on a
stable/icehouse neutron job, but for some reason
it doesn't show up in logstash in the gate or on
stable/icehouse.

Regardless, there are 82 hits in 10 days on multiple
changes, it affects pretty much any test that's
going through teardown with neutron and must be
latent to at least Icehouse.

I'm guessing this is due to the neutron full job
running with tenant isolation which causes multiple
ports to be using the same subnet and when one
test tries to delete the subnet it fails because
another test (port) is using the same subnet.

Change-Id: I03bba460f31f8769a843f7b77d6c5e684ca8ebf0
Related-Bug: #1357055