Add query for 1357476
The issue sounds similar to 1333654, but is in a
Fix Released status.

Now almost all match caused Failure 32 in 7 days,
only 1 match from icehouse was successful.
Because the 'check-tempest-dsvm-neutron-full-icehouse' is non
voting it will be excluded from the query.

VirtualInterfaceCreateException can happen even with
nova-network so the 'Timeout waiting for vif plugging callback for
instance' used in the query.

Change-Id: I59d02fee9580bf55dd5ad2786ca9a2a5e4b1bcb3
Related-Bug: #1357476
Related-Bug: #1333654