Make unit tests around bug_urls_map order indifferent
This commit fixes 2 unit tests which were assuming a specific dict
order after calling bug_urls_map(). When setting a random python
hashseed the tests would fail because the dict order changed. So the
assert checks after calling the method where changed to not assume
that the order was fixed.

Change-Id: Ie15e40cffd9511df4f8cf3f6d7f61eb4acced083