Update query for pip timeout bug 1270710
I noticed the uncategorized fail rates for the
gate-nova-python27 job were low and one of the
reasons is due to the pip timeout bug and the
existing query not hitting on new traces.

I really hate that we're updating this and
fingerprinting on line numbers in the packaging
code but I'm not aware of anything better at
the moment, so I guess this will just keep
being a moving target.

In 24 hours this does bring the hits up from
250 to 298.

Change-Id: I90ce79f4b268c036627b30fab03f9680eaa32304
Related-Bug: #1270710