Disambiguate filename:<file> from tag:<file> in query guidelines
It is not clear what is referred to by 'filename query' when it is
recommended in place of a 'console query'. One might think it is the
query of the form filename:<filename>. But it is then recommended to use
the 'tags' field rather than the 'filename'.

This commit removes the ambiguity.

Related-Bug: #1368877

Change-Id: I980584bf3afba681321318aecffb566bfcf57e8b