update web pages to support additional date info
This makes the web ui support some additional timestamps that will be
brought in from the json data. It's done in a compatible way so that
if this change lands before the data payload updates land it won't
break horribly.

This will let us display how far behind we are with our data
generation.

Change-Id: If4926879569a36cdd29d7e261ef556ef1a64b39d