Add query for lvm hang bug 1373513
111 hits in 7 days, check and gate.

This isn't a 100% failure rate, but when we see this
it's a failed job 98% of the time, so it's pretty close.

The tempest volume delete timeout is 196 seconds and we've
seen that pass by just a few seconds of seeing the volume
actually deleted in the cinder logs, so this is probably
the best query we have right now.

Change-Id: I280acc2c237d4ff17216b828062af3836b7ac3bf
Related-Bug: #1373513