Remove old queries for fixed or incomplete bugs
Most of these are fixed, or are duplicates of bugs
that are fixed.

Some of these are old incomplete bugs that we're
not sure are still issues, so we remove the query
and if the problem shows up again we'll restore
the query.

We also have a lot of queries for new/confirmed
bugs which might have changed fingerprints so we
need to update those, I've left them so we can
track them and do the updates as needed.

Change-Id: Ia51376f715c55db88171a0f51605d8e59f41fba2