Update query for volume delete hang bug 1373513
Now that we have better logging in tempest via
commit d2b9651cb82910b2b86b0f94af1721f717e74024,
we can update the query to include the volume
delete timeout message from the tempest logs.

This adds another 27 hits in the last 24 hours
to the failure count, which should get our
gate categorization percentage up from the measly
58% we have now.

As for the actual signatures, we're basically
treating bug 1373513 as the home for all of the
current cinder blocking lvm/lvs/vgs issues right
now. When we see the volume delete timeout message
in tempest we don't always see the lvm message
in syslog, so it could be problems with vgs taking
too long also (we know that's an intermittent
problem), but we suspect the root issues are the
same in Cinder and need to be investigated together.

Change-Id: Ia100f7c5174a9629a90df7b13030b88195baac12
Related-Bug: #1373513