Add query for tempest negative test bug 1392860
Looks like this just happened and it's hitting multiple
branches so I'm assuming it's a recent regression in
Tempest rather than a bug in Nova or Neutron.

25 hits in 24 hours, so far only listed against the check
queue but on multiple changes.

Filtered out the 'silent' queue since there were similar
but different hits there for the nova-docker job.

Change-Id: Ibd95aa1056c83b7c76720d213fce195df3186d5b
Related-Bug: #1392860