Add query for neutron port mismatch bug 1407190
348 hits in 7 days, check and gate, all failures.
This is a MismatchError in the tempest code so
we need a better error message in tempest when
the assertion fails to have a more specific
fingerprint, that will be worked separately
and then the query can be updated.

Also, it looks like this started on 12/29 and there
are several neutron commits merged on 12/29 that
look suspect for regression related to port/
interface changes.

Change-Id: Ia14da1578036fbdefec4c7fdeb694bbc17807aa5
Related-Bug: #1407190