Expand failure pattern for bug 1402747
It looks like the version in the error on juno are different than master
This commit makes the first half of the query version independent to
detect the same failure on all the branches. Doing this also makes the
second half of the query unnecessary since it will be also detected by
the console failure.

Co-Authored-By: Joe Gordon <joe.gordon0@gmail.com>
Change-Id: I4ff352554a83a9acc79c95a6fdfd7b5a6734dc86