Add query for nova unit test race bug 1403564
105 hits in 9 days, check and gate, all fails,
multiple changes.

Restricting to just gate queue since it's a unit
test, but still 42 hits in 9 days in the gate.

When restricted to the gate queue, it's only hitting
on one change but the change is pretty much unrelated
to what's failing here, so this must just be very
intermittent.

Change-Id: I4fa04aa3d88a44cf38e2ee4f5f1fa4628c2dde5b
Related-Bug: #1403564