Add query for boto 2.35 auth failure bug 1408987
3042 hits in 48 hours, check queue right now since
the boto version has been pinned, multiple changes
though, all failures.

This query is more specific since it's using a trace
in the tempest log, the hits are much higher if you
just query on the failure string in the console log
but those are a lot of duplicates.

Change-Id: If7771d66b61f66475100e6f6afee772e40922e50
Related-Bug: #1408987