Add query for heat unit test fail bug 1408704
This has 3080 hits in 7 days when you don't
filter on the gate queue, but since it's a unit
test failure I'm filtering on the gate queue,
and because it's fixed on master - so I just want
to get it off the uncategorized bugs page.

Otherwise 349 hits and it looks like something
needs to be fixed in heat on stable/juno yet
because this is still failing (locally for me).

Change-Id: If24be8a7ea330b4708f1679c7ba68e8c1e925f59
Related-Bug: #1408704