Add query for nova/neutron bug 1411312
9 hits in 7 days, check and gate, all failures.

There aren't any good messages in the nova/neutron
logs from what I can tell on this race failure,
so I'm fingerprinting on the error in the console
log. This should be relatively safe since the expected
IPs in the test are hard-coded and don't show up
elsewhere in Tempest.

Change-Id: Id9de33d5216103c1be884989410b68da4743e29d
Related-Bug: #1411312