Add query for django-openstack-auth 1.1.8 bug 1413876
55 hits in 7 days, check and gate, all failures.

This is already fixed as of today with the
django-openstack-auth 1.1.9 release, so this query
is to get it off the uncategorized bugs list.

We query on the specific library version in the pip
freeze output because otherwise there are hundreds
of different unit test failures to query so this is
the cleanest way to find the failures.

Change-Id: Id8674516b5c77cea6cbc795a556a3b1bb6f3da7c
Related-Bug: #1413876