Add query for neutron DBDeadlock bug 1426543
80 hits in 24 hours, check and gate, not all
failures since there is probably a retry on deadlock
in play here. Trends toward 76% failure when this
shows up though.

There is a related query in n-cpu for a connection
read timeout failure with neutronclient that has
a better failure rate (94%) but given the source of
the neutronclient read timeout is probably the
deadlock in the neutron DB, it's better to fingerprint
on the cause rather than the symptom.

Change-Id: I7e0a2199d1a9c561e2948ab95b508311a3b7ec9a
Related-Bug: #1426543