Remove old queries for "fixed" bugs
Removes queries that haven't hit in over 10 days
on bugs that are either fixed (some are duplicates
of fixed bugs), or are incomplete/invalid to the point
where they aren't issues anymore - or at least not seen
in a long time.  If they come back, we can write a new
query for a new bug.

Change-Id: I701c17fce69cb2a4a7f1aa5d985602783f799d12