add web_server.py for testing web content
Actually testing the web content for elastic recheck was a bit
annoying because it involved manually downloading json files and
modifying html files to point at a local server. This means that we
often pushed changes without testing them, sometimes with not so great
results.

This uses the stdlib BaseHTTPServer to build a fake local server that
runs on port 8001 and provides a local experience that pretends pretty
damn hard to be like hitting the live site. All local files are served
up statically. All references to upstream .json files are rewriten
into local server urls, and fetching of .json files that we don't have
in a local tree are automatically proxied out to live data on
status.openstack.org.

Change-Id: I4ae0fc67f8ab440153fc7edca5af77913cc448a7