Update query for bug 1403291
This is showing up in other traces and the tempest
logs seem to have derailed as of late, so generalize
this query to catch more hits.

I realize this is a generic failure and that sucks,
but the neutron logs are of no use here so we're kind
of stuck for the time being until we get better
logging to make this more specific.

66 hits in 10 days, multiple queues, all failures.

Change-Id: I6a8ec0ec6d1dbdae6592df02813c630901e23453