Remove duplicate query for bug 1466696
We already had a bug and a query for the cells
stop_instance InstanceIsLocked bug, I didn't notice
it because (1) launchpad couldn't tell the duplicate and
(2) elastic-recheck didn't report on my change that failed
with that bug, even though the query should hit (the queue
was probably backed up).

Anyway, we don't need two so remove the duplicate.

Change-Id: I5a55d981544cf67bbb213fba12e55996a76ce4cb