Add suppress-notification to the slow devstack query
This commit adds suppress-notification to the slow devstack query.
We've seen a large uptick in this being a hit lately, it's likely
because devstack has just gotten slightly slower again pushing us
over the wall more frequently. However, this message does not
directly any tests failures to the node being slow, it never really
did. So this will likely just mask real bugs as users think it's
not their problem. So let's still track when this happens on a graph
but not report to users that this is the likely cause for a failure
because it isn't.

Change-Id: Ic2a6c1403cab26403c854b452a8f1c7119bbb0d7